requirements.txt
## dependencies

# our custom configobj
https://code.rhodecode.com/upstream/configobj/artifacts/download/0-012de99a-b1e1-4f64-a5c0-07a98a41b324.tar.gz?md5=6a513f51fe04b2c18cf84c1395a7c626#egg=configobj==5.0.6

dogpile.cache==0.9.0
dogpile.core==0.4.1
decorator==4.1.2
dulwich==0.13.0
hgsubversion==1.9.3
hg-evolve==9.1.0
mako==1.1.0
markupsafe==1.1.1
mercurial==5.1.1
msgpack-python==0.5.6

pastedeploy==2.1.0
pyramid==1.10.4
pyramid-mako==1.1.0
pygit2==0.28.2

repoze.lru==0.7
redis==3.4.1
simplejson==3.16.0
subprocess32==3.5.4
subvertpy==0.10.1

six==1.11.0
translationstring==1.3
webob==1.8.5
zope.deprecation==4.4.0
zope.interface==4.6.0

## http servers
gevent==1.5.0
greenlet==0.4.15
gunicorn==19.9.0
waitress==1.3.1

## debug
ipdb==0.13.2
ipython==5.1.0

## test related requirements
-r requirements_test.txt

## uncomment to add the debug libraries
#-r requirements_debug.txt
